chore: minor rename
