+++ app/routes/decks/$id/history-graph.tsx
@@ -21,8 +21,8 @@ import { cls } from "../../../utils/misc";
 import type { PageHandle } from "../../../utils/page-handle";
 import { formatDateRange } from "../../../utils/temporal-utils";
 
-import type { LoaderData } from "./common.server";
-export { loader } from "./common.server";
+import type { LoaderData } from "./_utils.server";
+export { loader } from "./_utils.server";
 
 export const shouldRevalidate = disableUrlQueryRevalidation;
